+++ sdk/ml/azure-ai-ml/azure/ai/ml/_restclient/v2024_10_01_preview/models/_models.py
@@ -30369,6 +30369,10 @@ class Workspace(Resource):
      list[~azure.mgmt.machinelearningservices.models.PrivateEndpointConnection]
     :ivar private_link_count: Count of private connections in the workspace.
     :vartype private_link_count: int
+    :ivar provision_network_now: Set to trigger the provisioning of the managed VNet with the
+     default Options when creating a Workspace with the managed VNet enabled, or else it does
+     nothing.
+    :vartype provision_network_now: bool
     :ivar provisioning_state: The current deployment state of workspace resource. The
      provisioningState is to indicate states for resource provisioning. Possible values include:
      "Unknown", "Updating", "Creating", "Deleting", "Succeeded", "Failed", "Canceled".
@@ -30469,6 +30473,7 @@ class Workspace(Resource):
         'primary_user_assigned_identity': {'key': 'properties.primaryUserAssignedIdentity', 'type': 'str'},
         'private_endpoint_connections': {'key': 'properties.privateEndpointConnections', 'type': '[PrivateEndpointConnection]'},
         'private_link_count': {'key': 'properties.privateLinkCount', 'type': 'int'},
+        'provision_network_now': {'key': 'properties.provisionNetworkNow', 'type': 'bool'},
         'provisioning_state': {'key': 'properties.provisioningState', 'type': 'str'},
         'public_network_access': {'key': 'properties.publicNetworkAccess', 'type': 'str'},
         'serverless_compute_settings': {'key': 'properties.serverlessComputeSettings', 'type': 'ServerlessComputeSettings'},
@@ -30560,6 +30565,10 @@ class Workspace(Resource):
         :keyword primary_user_assigned_identity: The user assigned identity resource id that represents
          the workspace identity.
         :paramtype primary_user_assigned_identity: str
+        :keyword provision_network_now: Set to trigger the provisioning of the managed VNet with the
+         default Options when creating a Workspace with the managed VNet enabled, or else it does
+         nothing.
+        :paramtype provision_network_now: bool
         :keyword public_network_access: Whether requests from Public Network are allowed. Possible
          values include: "Enabled", "Disabled".
         :paramtype public_network_access: str or
@@ -30625,6 +30634,7 @@ class Workspace(Resource):
         self.primary_user_assigned_identity = kwargs.get('primary_user_assigned_identity', None)
         self.private_endpoint_connections = None
         self.private_link_count = None
+        self.provision_network_now = kwargs.get('provision_network_now', None)
         self.provisioning_state = None
         self.public_network_access = kwargs.get('public_network_access', None)
         self.serverless_compute_settings = kwargs.get('serverless_compute_settings', None)