+++ sdk/ml/azure-ai-ml/samples/ml_samples_workspace.py
@@ -16,9 +16,11 @@ USAGE:
 """
 
 import os
+
+from ml_samples_compute import handle_resource_exists_error
+
 from azure.ai.ml import MLClient
 from azure.identity import DefaultAzureCredential
-from ml_samples_compute import handle_resource_exists_error
 
 subscription_id = os.environ["AZURE_SUBSCRIPTION_ID"]
 resource_group = os.environ["RESOURCE_GROUP_NAME"]
@@ -75,12 +77,12 @@ class WorkspaceConfigurationOptions(object):
 
         # [START workspace_managed_network]
         from azure.ai.ml.entities import (
-            Workspace,
-            ManagedNetwork,
-            IsolationMode,
-            ServiceTagDestination,
-            PrivateEndpointDestination,
             FqdnDestination,
+            IsolationMode,
+            ManagedNetwork,
+            PrivateEndpointDestination,
+            ServiceTagDestination,
+            Workspace,
         )
 
         # Example private endpoint outbound to a blob
@@ -107,6 +109,13 @@ class WorkspaceConfigurationOptions(object):
         # Workspace configuration
         ws = Workspace(name="ws-name", location="eastus", managed_network=network)
         # [END workspace_managed_network]
+
+        # [START workspace_managed_network_provision_now]
+        from azure.ai.ml.entities import IsolationMode, ManagedNetwork, Workspace
+
+        managed_net = ManagedNetwork(isolation_mode=IsolationMode.ALLOW_INTERNET_OUTBOUND)
+        ws = Workspace(name="ws-name", location="eastus", managed_network=managed_net, provision_network_now=True)
+        # [END workspace_managed_network_provision_now]
 
         # [START fqdn_outboundrule]
         from azure.ai.ml.entities import FqdnDestination
@@ -230,8 +239,7 @@ class WorkspaceConfigurationOptions(object):
 
         # [START create_or_update_connection]
         from azure.ai.ml import MLClient
-        from azure.ai.ml.entities import WorkspaceConnection
-        from azure.ai.ml.entities import UsernamePasswordConfiguration
+        from azure.ai.ml.entities import UsernamePasswordConfiguration, WorkspaceConnection
 
         ml_client_ws = MLClient(credential, subscription_id, resource_group, workspace_name="test-ws")
         wps_connection = WorkspaceConnection(